+++ forge/src/main/java/fr/raksrinana/editsign/forge/config/cloth/ClothConfigHook.java
@@ -1,9 +1,18 @@
 package fr.raksrinana.editsign.forge.config.cloth;
 
+import fr.raksrinana.editsign.forge.config.CommonConfig;
+import fr.raksrinana.editsign.forge.config.Config;
 import lombok.NoArgsConstructor;
+import me.shedaniel.clothconfig2.api.ConfigBuilder;
+import me.shedaniel.clothconfig2.api.ConfigCategory;
+import me.shedaniel.clothconfig2.gui.entries.StringListEntry;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.TextComponent;
 import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
+import net.minecraftforge.fml.ModLoadingContext;
+import net.minecraftforge.fmlclient.ConfigGuiHandler.ConfigGuiFactory;
 import java.util.LinkedList;
 import java.util.Optional;
 import java.util.function.Function;
@@ -31,38 +40,38 @@ public class ClothConfigHook{
 	}
 	
 	public void load(){
-	// 	ModLoadingContext.get().registerExtensionPoint(ConfigGuiFactory.class, () -> new ConfigGuiFactory((minecraft, screen) -> {
-	// 		ConfigBuilder builder = ConfigBuilder.create()
-	// 				.setParentScreen(screen)
-	// 				.setTitle(new TextComponent("EditSign"));
-	//
-	// 		fillConfigScreen(builder);
-	//
-	// 		return builder.build();
-	// 	});
+		ModLoadingContext.get().registerExtensionPoint(ConfigGuiFactory.class, () -> new ConfigGuiFactory((minecraft, screen) -> {
+			ConfigBuilder builder = ConfigBuilder.create()
+					.setParentScreen(screen)
+					.setTitle(new TextComponent("EditSign"));
+			
+			fillConfigScreen(builder);
+			
+			return builder.build();
+		}));
 	}
 	
-	// @OnlyIn(Dist.CLIENT)
-	// public void fillConfigScreen(ConfigBuilder builder){
-	// 	CommonConfig config = Config.COMMON;
-	//
-	// 	StringListEntry reverseSneakingEntry = builder.entryBuilder()
-	// 			.startStrField(new TranslationTextComponent(getFieldName("requiredItemId")), config.getRequiredItemStr())
-	// 			.setDefaultValue("")
-	// 			.setTooltip(getTooltips("requiredItemId", 4))
-	// 			.setSaveConsumer(config::setRequiredItemId)
-	// 			.setErrorSupplier(getMinecraftItemIdCellError())
-	// 			.build();
-	//
-	// 	ConfigCategory general = builder.getOrCreateCategory(new TranslationTextComponent("text.autoconfig.editsign.category.default"));
-	// 	general.addEntry(reverseSneakingEntry);
-	// }
+	@OnlyIn(Dist.CLIENT)
+	public void fillConfigScreen(ConfigBuilder builder){
+		CommonConfig config = Config.COMMON;
+		
+		StringListEntry reverseSneakingEntry = builder.entryBuilder()
+				.startStrField(new TranslatableComponent(getFieldName("requiredItemId")), config.getRequiredItemStr())
+				.setDefaultValue("")
+				.setTooltip(getTooltips("requiredItemId", 4))
+				.setSaveConsumer(config::setRequiredItemId)
+				.setErrorSupplier(getMinecraftItemIdCellError())
+				.build();
+		
+		ConfigCategory general = builder.getOrCreateCategory(new TranslatableComponent("text.autoconfig.editsign.category.default"));
+		general.addEntry(reverseSneakingEntry);
+	}
 	
 	private String getFieldName(String fieldName){
 		return "text.autoconfig.editsign.option." + fieldName;
 	}
 	
-	private TextComponent[] getTooltips(String fieldName, int count){
+	private Component[] getTooltips(String fieldName, int count){
 		var tooltipKey = getFieldName(fieldName) + ".@Tooltip";
 		var keys = new LinkedList<String>();
 		if(count <= 1){
@@ -75,6 +84,6 @@ public class ClothConfigHook{
 		}
 		return keys.stream()
 				.map(TextComponent::new)
-				.toArray(TextComponent[]::new);
+				.toArray(Component[]::new);
 	}
 }